AIML Project - ML Model/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
joblib
pymongo
python-dotenv
requests
beautifulsoup4
schedule
